refactor: adjust ui theme folder structure
