+++ packages/cube-frontend-ui-theme/src/utils/animationKeyframes.ts
