+++ packages/cube-frontend-ui-theme/src/utils/animations.ts
